move minimizers to bundler package
