+++ packages/docusaurus-bundler/src/faster.ts
