+++ packages/docusaurus-bundler/src/minification.ts
@@ -7,13 +7,14 @@
 
 import TerserPlugin from 'terser-webpack-plugin';
 import CssMinimizerPlugin from 'css-minimizer-webpack-plugin';
-import {importSwcJsMinifierOptions} from '../faster';
+import {importSwcJsMinifierOptions} from './faster';
 import type {CustomOptions, CssNanoOptions} from 'css-minimizer-webpack-plugin';
 import type {WebpackPluginInstance} from 'webpack';
-import type {FasterConfig} from '@docusaurus/types';
+import type {CurrentBundler, FasterConfig} from '@docusaurus/types';
 
 export type MinimizersConfig = {
   faster: Pick<FasterConfig, 'swcJsMinimizer'>;
+  currentBundler: CurrentBundler;
 };
 
 // See https://github.com/webpack-contrib/terser-webpack-plugin#parallel
@@ -111,8 +112,23 @@ function getCssMinimizer(): WebpackPluginInstance {
   }
 }
 
-export async function getMinimizers(
+async function getWebpackMinimizers(
   params: MinimizersConfig,
 ): Promise<WebpackPluginInstance[]> {
   return Promise.all([getJsMinimizer(params), getCssMinimizer()]);
+}
+
+async function getRspackMinimizers({
+  currentBundler,
+}: MinimizersConfig): Promise<WebpackPluginInstance[]> {
+  console.log('currentBundler', currentBundler.name);
+  throw new Error('TODO Rspack minimizers not implemented yet');
+}
+
+export async function getMinimizers(
+  params: MinimizersConfig,
+): Promise<WebpackPluginInstance[]> {
+  return params.currentBundler.name === 'rspack'
+    ? getRspackMinimizers(params)
+    : getWebpackMinimizers(params);
 }
