+++ packages/docusaurus-plugin-pwa/package.json
@@ -33,7 +33,6 @@
     "babel-loader": "^9.1.3",
     "clsx": "^2.0.0",
     "core-js": "^3.31.1",
-    "terser-webpack-plugin": "^5.3.9",
     "tslib": "^2.6.0",
     "webpack": "^5.88.1",
     "webpack-merge": "^5.9.0",
