+++ packages/docusaurus/src/webpack/base.ts
@@ -7,10 +7,9 @@
 
 import fs from 'fs-extra';
 import path from 'path';
-import {getCSSExtractPlugin} from '@docusaurus/bundler';
+import {getCSSExtractPlugin, getMinimizers} from '@docusaurus/bundler';
 import {md5Hash, getFileLoaderUtils} from '@docusaurus/utils';
 import {createJsLoaderFactory, getCustomBabelConfigFilePath} from './utils';
-import {getMinimizers} from './minification';
 import {loadThemeAliases, loadDocusaurusAliases} from './aliases';
 import type {Configuration} from 'webpack';
 import type {
@@ -180,7 +179,9 @@ export async function createBaseConfig({
       // Only minimize client bundle in production because server bundle is only
       // used for static site generation
       minimize: minimizeEnabled,
-      minimizer: minimizeEnabled ? await getMinimizers({faster}) : undefined,
+      minimizer: minimizeEnabled
+        ? await getMinimizers({faster, currentBundler: props.currentBundler})
+        : undefined,
       splitChunks: isServer
         ? false
         : {
